feat(voice): add audio.cpp provider with local and remote GPU pool support
Implement AudioCppProvider in crates/movie-radio-voice supporting local
CLI/server and remote GPU pool HTTP inference endpoints with fallback
and budget policies. Modularize audio_cpp into sub-files to comply
with MAX_LINES_PER_SOURCE_FILE quality gates and code health.
